Add new components, refactor existing styles
Added new components Icon, ContactsSection, ProfileSection and PhotoSection, and updated styles files. Additional changes include file mode updates and modifications to existing components such as UserSection and Divider. These changes were necessary in order to improve the structure of the codebase and make it easier to manage styles as well as enhancing the maintainability of the codebase. By providing a more modular structure, these changes contribute to more versatile and reusable components.