Update font properties and site-wide styling in ResumeReview section
The styling properties of ResumeReview section were updated and site-wide styling was enhanced, leading to better readability and reusability of components. Font properties were changed to use Montserrat font, with different weights and sizes applied to different parts of the UserSection according to the new design guidelines. Importing of the new font from Google Fonts and general box-sizing attribute were added at the root level. This update is intended to improve the overall aesthetic and consistency of the website.