Integrate react-router-dom and refactor navigation
Integrate react-router-dom for navigation instead of custom context-based routing. Refactor all the components and pages where navigation is used to utilize the capabilities of react-router-dom. This change simplifies and unifies the way of controlling routes, enhances the application's scalability and maintainability, and provides better user experience with built-in browser navigation support. Also, in this refactor, provided a label prop in "Input" component to have custom labels instead of using names as labels. In addition, a sample "resumePreviewData" constant is added as a template to show user data in the "ResumePreview" component.